Clean up redundant type conversions and suppress deprecation warnings
- **Color & Palette Logic**:
    - Remove unnecessary `.toDouble()` calls in `PaletteFromHueExt.kt` and `SystemPalettesExt.kt` when calculating HCT values and contrast levels.
    - Simplify fraction calculation in `SystemPalettesExt.kt` by removing explicit double conversion.
- **Coroutines & Flows**:
    - Add `@Suppress("DEPRECATION")` to the `mapReusable` extension function in `StateFlowReusableExt.kt`.
- **Utilities**:
    - Rename parameters from `entry1`/`entry2` to `a`/`b` in the `ScoredComparator` implementation within `Score.kt`.

+++ app/model/src/androidMain/kotlin/org/hnau/commons/app/model/app/SystemPalettesExt.kt
@@ -182,11 +182,11 @@ private class SystemPalette(
         }
         val (fromTone, fromColor) = this[insertionPoint - 1]
         val (toTone, toColor) = this[insertionPoint]
-        val fraction = (tone.raw - fromTone.raw).toDouble() / (toTone.raw - fromTone.raw)
+        val fraction = (tone.raw - fromTone.raw) / (toTone.raw - fromTone.raw)
         return Hct.from(
             hue = lerp(fromColor.hue, toColor.hue, fraction),
             chroma = lerp(fromColor.chroma, toColor.chroma, fraction),
-            tone = tone.raw.toDouble(),
+            tone = tone.raw,
         )
     }
 }
